requirements.txt
discord.py
python-dotenv
fuzzysearch
discord-ext-voice-recv